% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_project_handling.R
\name{aggregate_depth}
\alias{aggregate_depth}
\title{Aggregate soilgrids Layers over Depth.}
\arguments{
\item{lower_bound}{Vector defining the lower bounds of the aggregated soil
layers (depths in cm).}
}
\description{
Aggregate the \href{https://soilgrids.org/}{soilgrids} layers within the
 \strong{soil project} over their respective depth. By convention most
 layers are divided into 7 depth-classes.This function allows to aggregated
 them over chosen depths.
}
