% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{calculate_max_dist}
\alias{calculate_max_dist}
\title{Calculate maximum difference for each cluster value to the chord line}
\usage{
calculate_max_dist(soil_data)
}
\arguments{
\item{soil_data}{Saved soil information in the soil_object.}
}
\description{
Calculate maximum difference for each cluster value to the chord line
}
\keyword{internal}
