% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_soilproperty.R
\name{calculate_soilproperty}
\alias{calculate_soilproperty}
\title{Calculate new soil properties.}
\usage{
calculate_soilproperty(soil_list, sl = NULL, ...)
}
\arguments{
\item{soil_list}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{sl}{Vector defining the soil layers for which the properties are calculated.}
}
