% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{cluster_soil}
\alias{cluster_soil}
\title{Cluster Layers}
\usage{
cluster_soil(soil_data, clusters_k)
}
\arguments{
\item{soil_data}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{clusters_k}{Vector of number of soil classes that should be generated
with k-means clustering.}
}
\description{
Cluster the layers within the \strong{soil project} by using kmeans
  clustering.
}
