% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cluster.R
\name{evaluate_cluster}
\alias{evaluate_cluster}
\title{Evaluate the cluster result calculating the sum of squared differences for
each clustering}
\usage{
evaluate_cluster(cluster_result)
}
\arguments{
\item{cluster_result}{Results of the soilgrids clustering stored in soil_object}
}
\value{
A ggplot object that shows the the SSE (within a the given classes)
  over the number of classes
}
\description{
Evaluate the cluster result calculating the sum of squared differences for
each clustering
}
