% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layermeta.R
\name{get_layermeta}
\alias{get_layermeta}
\title{get the required soilgrids}
\usage{
get_layermeta(wcs = "http://data.isric.org/geoserver/sg250m/wcs?",
  raw = FALSE)
}
\arguments{
\item{wcs}{URL of the ISRIC soilgrids geoserver}

\item{raw}{logic, if TRUE raw text output is returned if FALSE (default) the
function returns the extracted values of the Tiff tiles pixel sizen and the
extent.}
}
\value{
Returns either a list with pixel size and extent of Tiffs as numerics
  or the raw text output.
}
\description{
get the required soilgrids
}

