% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_project_handling.R
\name{load_project}
\alias{load_project}
\title{Load an existing Soil Project}
\usage{
load_project(soil_project)
}
\arguments{
\item{soil_project}{Character string providing path to the soil project.}
}
\value{
The loaded \code{\link{soil_project}} in the working environment (as an R6
  object).
}
\description{
Loads an existing \strong{soil project} (see: \code{\link{soil_project}}),
  together with a fixed link to the project folder and project path
  (used to save the data).
}
