% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_project_handling.R
\name{load_soil_project}
\alias{load_soil_project}
\title{Load an existing Soil Project}
\usage{
load_soil_project(soil_project)
}
\arguments{
\item{project_name}{The name of the soil project. This will be the name of
the project folder and the soil project in the working environment.
Once initialized \strong{the name should not be changed}!}

\item{project_path}{Path to the SWAT project / Path where soilgrids layers
are saved. Once initialized \strong{the path should not be changed}!}

\item{shape_file}{Shape file (or path to shape file) that defines the extent
of the soilgrids layers. Alternatively, the extend (\code{extend}) and
coordinate-system (\code{crs}) of a spatial object can be used.}
}
\value{
New \code{\link{soil_project}} in the working environment (as an R6
  object), as well as the project folder in the project path.
}
\description{
Loads an existing \strong{soil project} (see: \code{\link{soil_project}}),
  together with a fixed link to the project folder and project path
  (used to save the data).
}
