% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_soilgrids.R
\name{obtain_soilgrids}
\alias{obtain_soilgrids}
\title{Obtain Soilgrids}
\usage{
obtain_soilgrids(project_path, shp_file, variables, depths, quantiles, wcs)
}
\arguments{
\item{project_path}{Path to the SWAT project / Path where soilgrids layers
are saved}

\item{shp_file}{Shape file (or path to shape file) that defines the extent of
the soilgrids layers. If \code{NULL} the shape file from the SWAT project
will be used.}

\item{wcs}{URL of the ISRIC soilgrids geoserver}

\item{layer_meta}{List with the entries extent (Vector of length 4 with the
soilgrids extent) and the soilgrids pixel_size. Check if these values are
correct with \code{get_layermeta()]}}

\item{layer_names}{Character vector of names of the soilgrids layers to be
downloadad.}
}
\value{
Writes the required soilgrids layer to project_path/soilgrids.
}
\description{
Obtain soilgrids layer from the ISRIC geoserver.
}
\keyword{internal}
