% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustersoil.R
\name{plot_clustersoil}
\alias{plot_clustersoil}
\title{Plot clustered soil map}
\usage{
plot_clustersoil(soilgrids_cluster, n_class)
}
\arguments{
\item{n_class}{}
}
\description{
Plot clustered soil map
}

