% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{plot_soilmap}
\alias{plot_soilmap}
\title{Plot clustered soil map}
\usage{
plot_soilmap(soil_data, cluster_k)
}
\arguments{
\item{soil_data}{The cluster results output from \code{cluster_soilgrids()}.}

\item{cluster_k}{Numeric. Number of soil classes to plot in the soil map.}
}
\value{
Plots the final soilmap with the selected number of soil classes.
}
\description{
Internal/Actual general-function to plot the clustered soil maps.
}
\keyword{internal}
