% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot soil variables from processed data}
\usage{
plot_variable(soil_data, variable, sl)
}
\arguments{
\item{soil_data}{The cluster results output from \code{cluster_soilgrids()}.}

\item{variable}{Charcacter vector providing variable names.}

\item{sl}{Vector providing soil layers to be plotted}
}
\value{
Plots the final soilmap with the selected number of soil classes.
}
\description{
Internal/Actual general-function to plot the clustered soil maps.
}
\keyword{internal}
