% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plural}
\alias{plural}
\title{Add plural 's' to the written message if multiple operations done}
\usage{
plural(n)
}
\arguments{
\item{n}{Interger number of operations}
}
\description{
Add plural 's' to the written message if multiple operations done
}
\keyword{internal}
