% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_output.R
\name{return_out}
\alias{return_out}
\title{Write Soil Project Output}
\usage{
return_out(soil_data, variable, sl, format)
}
\arguments{
\item{soil_data}{The cluster results output from \code{cluster_soilgrids()}.}

\item{variable}{Charcacter vector providing variable names.}

\item{sl}{Vector providing soil layers to be plotted}

\item{format}{Output format to be returned (either tibble or raster).}
}
\value{
Writes the modified soilgrids layers in the defined format.
}
\description{
Internal/Actual soilgrids-writing function.
}
\keyword{internal}
