% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_project_handling.R
\name{select_n_class}
\alias{select_n_class}
\title{Select Number of Clusters}
\arguments{
\item{n}{Number of clusters.}
}
\description{
Select the number of clusters (within kmeans) for further processing.
The function is meant to be used in conjunction with
\code{\link{cluster_soil}} and \code{\link{evaluate_cluster}}.
}
