% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilproperty_modification.R
\name{select_soilproperty}
\alias{select_soilproperty}
\title{Select soil properties.}
\usage{
select_soilproperty(soil_list, sl = NULL, ...)
}
\arguments{
\item{soil_list}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{sl}{Vector defining the soil layers for which the properties are calculated.}

\item{...}{Further arguments provide soil porperties to select in \code{dplyr::select} style.}
}
\description{
Select soil properties provided in the \strong{soil projects'} soil layer tables.
}
