% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{set_cluster_data}
\alias{set_cluster_data}
\title{Set Final Cluster Number}
\usage{
set_cluster_data(soil_data, cluster_k)
}
\arguments{
\item{soil_data}{Saved soil information in the soil_object.}

\item{cluster_k}{Numeric value of final number of classes.}
}
\description{
Define the final number of clusters within the \strong{soil project} and
set the respective clustered soil data.
}
\keyword{internal}
