% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_depth_from_vrt.R
\name{soil_depth_from_vrt}
\alias{soil_depth_from_vrt}
\title{Getting the SoilGrids250m (2017) "Absolute Depth To Bedrock" (BDTICM)
layer through WebDAV, cropped to a reference raster's extent.}
\usage{
soil_depth_from_vrt(
  x,
  destdir = tempdir(),
  sg_url =
    "/vsicurl/https://files.isric.org/soilgrids/former/2017-03-10/data/BDTICM_M_250m_ll.tif",
  method = "near",
  align = TRUE,
  use_crop = TRUE,
  filename = NULL,
  overwrite = NA,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SpatRaster-class}} object, used only to define
the crop extent (its own values/resolution are not used, unlike
in \code{align = TRUE}, see below).}

\item{destdir}{directory where to write the cropped raw GeoTIFF.}

\item{sg_url}{URL (vsicurl) to the BDTICM GeoTIFF. See default.}

\item{method, align, use_crop, filename, overwrite, ...}{same meaning as in
\code{soilgridr::soilgridmap_from_vrt}: if \code{align = TRUE} the
output is reprojected/aligned onto \code{x}'s grid via
\code{\link{project}} (and optionally cropped to \code{x}'s exact
extent if \code{use_crop = TRUE}).}
}
\description{
Getting the SoilGrids250m (2017) "Absolute Depth To Bedrock" (BDTICM)
layer through WebDAV, cropped to a reference raster's extent.
}
\details{
Unlike the current SoilGrids VRTs (which are natively in
  Homolosine projection), BDTICM_M_250m_ll.tif is already in
  geographic coordinates (EPSG:4326, hence the "_ll" suffix), so no
  reprojection of the bounding box is required before cropping.
}
\examples{
library(terra)
f <- system.file('ex/elev_vinschgau.tif', package="terra")
r <- rast(f)
depth_default <- soil_depth_from_vrt(r)

}
\seealso{
\url{https://data.isric.org/geonetwork/srv/api/records/f36117ea-9be5-4afd-bb7d-7a3e77bf392a},
\url{https://files.isric.org/soilgrids/former/2017-03-10/data/}
}
