% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_object.R
\docType{data}
\name{soil_project}
\alias{soil_project}
\title{Soil Project Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
soil_project
}
\description{
Basic \code{solACE} building block
 (see: \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{Introduction to R6}). \cr
 Generate a soil project with the \code{\link{new_soil_project}} function.
}
\keyword{datasets}
