% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgridmap.R
\name{soilgridmap_from_vrt}
\alias{soilgridmap_from_vrt}
\title{Getting SoilGRid Dataset through WebDav: R procedure to download vrt fils and import into 'SpatRast' object.}
\usage{
soilgridmap_from_vrt(
  x,
  destdir = tempdir(),
  voi = c("bdod", "cec", "cfvo", "clay", "nitrogen", "phh2o", "sand", "silt", "soc",
    "ocd", "ocs")[4],
  zl = c(0, 5, 15, 30, 60, 100, 200),
  depths = sprintf("\%d-\%dcm", zl[-length(zl)], zl[-1]),
  ndepths = sprintf("\%03d-\%03dcm", zl[-length(zl)], zl[-1]),
  quantiles = c("Q0.05", "Q0.50", "mean", "Q0.95", "Uncertainty")[c(-2, -5)],
  igh = "+proj=igh +lat_0=0 +lon_0=0 +datum=WGS84 +units=m +no_defs",
  sg_url =
    "/vsicurl?max_retry=3&retry_delay=1&list_dir=no&url=https://files.isric.org/soilgrids/latest/data/",
  method = "near",
  align = FALSE,
  use_crop = FALSE,
  filename = NULL,
  overwrite = NA,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SpatRaster-class}} object.}

\item{destdir}{directory where to write the SoilGrids output raw files.}

\item{voi}{variable of interest. See options.}

\item{zl}{soil layer depths (numeric)}

\item{depths, ndepths}{soil layer names as named in SoilGrid and written in a 3-digit form. See default.}

\item{quantiles}{quantile names. See default.}

\item{igh}{proj string for Homolosine projection. See SoilGRid documentation (URL below).}

\item{sg_url}{url prefix (directory) to vrt files.  See SoilGRid documentation (URL below).}

\item{method, align, filename, overwrite, ...}{further arguments for \code{\link{project}} (and \code{\link{writeRaster}}) in which its argument \code{y} is set equal to this function's argument \code{x}.}

\item{use_crop}{logical. Default is \code{FALSE}. If \code{TRUE}, the outcome is cropped with \code{x} extent.}
}
\description{
Getting SoilGRid Dataset through WebDav: R procedure to download vrt fils and import into 'SpatRast' object.
}
\examples{

library(terra)
f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
soilgrid_default_clay <- soilgridmap_from_vrt(r)


library(stars)
olinda <-  rast(system.file('tif/olinda_dem_utm25s.tif',package="stars"))
olinda_soilgrid_default_sand <- soilgridmap_from_vrt(olinda,voi="sand",align=TRUE)

vinschgau <-  rast(system.file('ext_data/vinschgau_elevation.tif',package="soilgridr"))
vinschgau_soilgrid_default_sand <- soilgridmap_from_vrt(vinschgau,voi="sand",align=TRUE,use_crop=TRUE)




}
\seealso{
\url{https://git.wur.nl/isric/soilgrids/soilgrids.notebooks/-/blob/master/markdown/webdav_from_R.md},
\url{https://www.isric.org/explore/soilgrids/faq-soilgrids} ,
\url{https://essd.copernicus.org/articles/12/299/2020/essd-12-299-2020.html},
\url{https://www.isric.org/sites/default/files/GlobalSoilMap_specifications_december_2015_2.pdf}
}
