% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_out}
\alias{write_out}
\title{Write Soil Project Output}
\usage{
write_out(soil_data, format, overwrite = FALSE)
}
\arguments{
\item{soil_data}{Soil data.}

\item{format}{Output format provided to the \link[rgdal]{writeGDAL} function (e.g. tif, ascii, etc.).}

\item{overwrite}{Logical. Defines if potentially existing outputs should overwritten or not.}
}
\value{
Writes the modified soilgrids layers in the defined format.
}
\description{
Internal/Actual soilgrids-writing function.
}
\keyword{internal}
