% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write the modified soilgrids layer}
\usage{
write_output(soil_data, format, overwrite = FALSE)
}
\arguments{
\item{soil_data}{soil data.}

\item{format}{tif, ascii etc...}

\item{overwrite}{Logical}
}
\value{
Writes the modified soilgrids layers in the defined format
}
\description{
Write the modified soilgrids layer
}
