% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_layer.R
\name{aggregate_layer}
\alias{aggregate_layer}
\title{Aggregate soilGrids Layers over Depth.}
\usage{
aggregate_layer(soil_list, lower_bound)
}
\arguments{
\item{soil_list}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{lower_bound}{Vector defining the lower bounds of the aggregated soil
layers (depths in cm).}
}
