% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilproperty_modification.R
\name{calculate_soilproperty}
\alias{calculate_soilproperty}
\title{Calculate new soil properties.}
\usage{
calculate_soilproperty(soil_list, sl = NULL, fun_list)
}
\arguments{
\item{soil_list}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{sl}{Vector defining the soil layers for which the properties are calculated.}

\item{...}{Further arguments provide functions to calculate new soil porperties in \code{dplyr::mutate} style.}
}
\description{
Calculate new soil properties from the \href{https://soilgrids.org/}{soilgrids}
layers within the \strong{soil project}
}
