% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soilgrids.R
\name{check_soilgrids}
\alias{check_soilgrids}
\title{Check the soilgrids meta data}
\usage{
check_soilgrids(wcs = "http://data.isric.org/geoserver/sg250m/wcs?")
}
\arguments{
\item{wcs}{URL of the ISRIC soilgrids geoserver}
}
\description{
Check the soilgrids meta data
}
