% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{cluster_soil}
\alias{cluster_soil}
\title{Cluster soilgrids layers using kmeans clustering}
\usage{
cluster_soil(soil_list, n_class)
}
\arguments{
\item{soil_list}{List of tibbles holding soil parameters for the respective
soil layers.}

\item{n_class}{Vector of number of soil classes that should be generated
with k-means clustering.}
}
\description{
Cluster soilgrids layers using kmeans clustering
}
