% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_soil.R
\name{evaluate_cluster}
\alias{evaluate_cluster}
\title{Evaluate Clustering Results}
\usage{
evaluate_cluster(soil_data)
}
\arguments{
\item{cluster_result}{Results of the soilgrids clustering stored in the \strong{soil project}.}
}
\value{
A ggplot object that shows the the SSE (within a the given classes)
  over the number of classes
}
\description{
Evaluate the results of the different \link[=cluster_soil]{kmeans clusters} by calculating the
  sum of squared differences for each one (between the clustered and the raw data).
}
