% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cluster.R
\name{evaluate_cluster}
\alias{evaluate_cluster}
\title{Plotting the cluster results for decision for number of soil classes}
\usage{
evaluate_cluster(sg_cluster)
}
\arguments{
\item{sg_cluster}{Results of the soilgrids clustering}
}
\value{
Returns a ggplot object plotting the within class SSE over the
  Number of classes
}
\description{
Plotting the cluster results for decision for number of soil classes
}

