% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cluster.R
\name{evaluate_cluster}
\alias{evaluate_cluster}
\title{Plotting of the cluster results}
\usage{
evaluate_cluster(cluster_result)
}
\arguments{
\item{sg_cluster}{Results of the soilgrids clustering}
}
\value{
A ggplot object that shows the the SSE (within a the given classes) over the
  number of classes
}
\description{
Creates a plot of the results from the soilgrids clustering (\code{sg_cluster}).
}

