% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layermeta.R
\name{get_layermeta}
\alias{get_layermeta}
\title{Soilgrids Metainformation}
\usage{
get_layermeta(project_path,
  wcs = "http://data.isric.org/geoserver/sg250m/wcs?", raw = FALSE)
}
\arguments{
\item{wcs}{URL of the ISRIC soilgrids geoserver}

\item{raw}{logic, if \code{TRUE}  raw text output is returned if \code{FALSE} (default) the
function returns the extracted values of the Tiff tiles pixel size and the
extent.}
}
\value{
Returns either a list with pixel size and extent of Tiffs as numerics
  or the raw text output.
}
\description{
Get the required soilgrids meta information.
}
\examples{
# Generate the required input for `obtain_soilgrids()`:
  layer_meta <- get_layermeta()
  layer_meta

# Check all meta information:
# (also for crosschecking that function works properly)
  get_layermeta(raw = TRUE)
}

