% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layermeta.R
\name{get_layermeta}
\alias{get_layermeta}
\title{Soilgrids Metainformation}
\usage{
get_layermeta(project_path, wcs)
}
\arguments{
\item{project_path}{Path where the soil_project is located}

\item{wcs}{URL of the ISRIC soilgrids geoserver}
}
\value{
Returns either a list with pixel size and extent of Tiffs as numerics
  or the raw text output.
}
\description{
Get the required soilgrids meta information.
}
\examples{
# Generate the required input for `obtain_soilgrids()`:
  layer_meta <- get_layermeta()
  layer_meta

# Check all meta information:
# (also for crosschecking that function works properly)
  get_layermeta(raw = TRUE)
}
