% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soilgrids.R
\name{load_soilgrids}
\alias{load_soilgrids}
\title{Load soilgrids layers into soil object}
\usage{
load_soilgrids(project_path, shape_file, layer_names)
}
\arguments{
\item{project_path}{Path to the soil project}

\item{shape_file}{Shape file of the study area}

\item{layer_names}{Character vector defining the soilgrids layers to load}
}
\description{
Load soilgrids layers into soil object
}
