% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soilgrids.R
\name{load_soilgrids}
\alias{load_soilgrids}
\title{Load soilgrids}
\usage{
load_soilgrids(project_path, shape_file, layer_table)
}
\arguments{
\item{project_path}{Path to the soil project.}

\item{shape_file}{Shape file of the study area.}

\item{layer_names}{Character vector for defining the soilgrids layers to load, e.g "BDRICM_M_250m".}
}
\description{
Loads a set of layers (defined by \code{layer_names} character-vector) from
 \href{https://soilgrids.org/}{soilgrids} into the \strong{soil project}
 (this includes files in the defined path and the R environment, see:
 \code{\link{new_soil_project}}).
}
\keyword{internal}
