% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_soil_project.R
\name{new_soil_project}
\alias{new_soil_project}
\title{Initiate a new soil project}
\usage{
new_soil_project(project_name, project_path, shape_file = NULL,
  extent = NULL, crs = NULL)
}
\arguments{
\item{project_name}{the name of the soil project. This will be the name of
the project folder and the soil project in the working environment. The
name of this object should not be changed!}

\item{project_path}{Path to the SWAT project / Path where soilgrids layers
are saved}

\item{shape_file}{Shape file (or path to shape file) that defines the extent
of the soilgrids layers.}
}
\value{
Generates a new soil_project in the working environment (as an R6
  object) and generates the prject folder in the project path.
}
\description{
Initiate a new soil project
}

