% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_soilgrids.R
\name{obtain_soilgrids}
\alias{obtain_soilgrids}
\title{Obtain soilgrids layer from the ISRIC geoserver}
\usage{
obtain_soilgrids(project_path, shp_file = NULL,
  wcs = "http://data.isric.org/geoserver/sg250m/wcs?", sg_ext = c(-180, 180,
  -56.0008104, 83.9991672), sg_pxl = 1/480)
}
\arguments{
\item{project_path}{Path to the SWAT project / Path where soilgrids layers
are saved}

\item{shp_file}{Shape file (or path to shape file) to define the extent of
the soilgrids layers. If \code{NULL} the shape file from the SWAT project
will be used.}

\item{wcs}{URL of the ISRIC soilgrids geoserver}

\item{sg_ext}{Extent of the soilgrids layers at the geoserver}

\item{sg_pxl}{Pixel size of the soilgrids layers at the geoserver. Check
default values of with \code{check_soilgrids()}}
}
\value{
Writes the requiered soilgrids layer to project_path/soilgrids.
}
\description{
Obtain soilgrids layer from the ISRIC geoserver
}
