% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_soilgrids.R
\name{obtain_soilgrids}
\alias{obtain_soilgrids}
\title{Obtain soilgrids layer from the ISRIC geoserver}
\usage{
obtain_soilgrids(project_path, shp_file = NULL,
  wcs = "http://data.isric.org/geoserver/sg250m/wcs?",
  layer_meta = list(extent = c(-180, 180, -56.0008104, 83.9991672), pixel_size
  = 1/480))
}
\arguments{
\item{project_path}{Path to the SWAT project / Path where soilgrids layers
are saved}

\item{shp_file}{Shape file (or path to shape file) that defines the extent of
the soilgrids layers. If \code{NULL} the shape file from the SWAT project
will be used.}

\item{wcs}{URL of the ISRIC soilgrids geoserver}

\item{layer_meta}{List with the entries extent (Vector of length 4 with the
soilgrids extent) and the soilgrids pixel_size. Check if these values are
correct with \code{get_layermeta()]}}
}
\value{
Writes the required soilgrids layer to project_path/soilgrids.
}
\description{
Obtain soilgrids layer from the ISRIC geoserver
}

