% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cluster_data.R
\name{set_cluster_data}
\alias{set_cluster_data}
\title{Define the final number of clusters and set the respective clustered soil data}
\usage{
set_cluster_data(soil_data, n_class)
}
\arguments{
\item{soil_data}{saved soil information in the soil_object}

\item{n_class}{Numeric value of final number of classes.}
}
\description{
Define the final number of clusters and set the respective clustered soil data
}
