% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_object.R
\docType{data}
\name{soil_project}
\alias{soil_project}
\title{soil_project}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
soil_project
}
\description{
soil_project
}
\keyword{datasets}
