% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_object.R
\docType{data}
\name{soil_project}
\alias{soil_project}
\title{Soil Project Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
soil_project
}
\description{
The \strong{soil project} class constitutes the basic building block of
  \code{solACE}. One can Generate a soil project with the
  \code{\link{new_project}} function. \cr
  In case you are interested in the technical details: The soil
  project class is realised by using object-oriented approach from the R6
  package. Read more at: \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{Introduction to R6}.
}
\keyword{datasets}
