% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_object.R
\name{soil_project}
\alias{soil_project}
\title{Soil Project Class}
\description{
The \strong{soil project} class constitutes the basic building block of
  \code{solACE}. One can Generate a soil project with the
  \code{\link{new_project}} function. \cr
  In case you are interested in the technical details: The soil
  project class is realised by using object-oriented approach from the R6
  package. Read more at: \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{Introduction to R6}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{soil_project$new()}}
\item \href{#method-print}{\code{soil_project$print()}}
\item \href{#method-save}{\code{soil_project$save()}}
\item \href{#method-load_soilgrids}{\code{soil_project$load_soilgrids()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{soil_project$new(project_name, project_path, shape_file, ext, crs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{soil_project$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{soil_project$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_soilgrids"></a>}}
\subsection{Method \code{load_soilgrids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{soil_project$load_soilgrids(
  variables = c("sand", "silt", "clay", "bdod", "cfvo", "cec", "phh2o", "soc"),
  depths = 1:6,
  quantiles = "mean"
)}\if{html}{\out{</div>}}
}

}
}
