% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_texture_class.R
\name{soil_texture_class}
\alias{soil_texture_class}
\title{Soil textura classes}
\usage{
soil_texture_class(x = NULL, sand = "sand", clay = "clay", silt = "silt", ...)
}
\arguments{
\item{x}{input soil maps}
}
\description{
Soil textura classes
}
\examples{

library(terra)
library(stars)
olinda <-  rast(system.file('tif/olinda_dem_utm25s.tif',package="stars"))
olinda_soilgrid_default_sand <- soilgridmap_from_vrt(olinda,voi="sand",align=TRUE,use_crop=TRUE)



}
