% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write the required SWAT2012 soil inputs for clustered soilgrids}
\usage{
write_output(soil_data, format, overwrite)
}
\arguments{
\item{soil_data}{soil data.}

\item{format}{tif, ascii etc...}

\item{overwrite}{Logical}
}
\value{
Writes required aggregated SWAT soil data to folder project_path/soil_out
}
\description{
Write the required SWAT2012 soil inputs for clustered soilgrids
}

