% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SWATsoil.R
\name{write_SWATsoil}
\alias{write_SWATsoil}
\title{write the required SWAT2012 soil inputs for clustered soilgrids}
\usage{
write_SWATsoil(project_path, sg_cluster, n_class, overwrite = FALSE)
}
\arguments{
\item{n_class}{}
}
\description{
write the required SWAT2012 soil inputs for clustered soilgrids
}
